pub/sub sync example: move the sleep to the subscriber side
This more closely matches the C code in the zguide example. The
paragraph below the example explains the reason for the "brute
force sleep" of one second between subscribing and sending the
synchronization: we can't assume that the SUB connect will be
finished by the time the sync REQ/REP dialog is complete.