improved process killing when running examples
When I run the examples, I have a tmux split open with a shell loop
showing me what example processes are running. Mostly, the examples stop
after the timeout has elapsed, but before this commit, I was
occasionally seeing a few processes continue to hang around after the
`boot examples -a` process exited.

I think `sh/exit-code` doesn't 100% get the job done with respect to
killing the process after the timeout has elapsed. I'm not sure why this
is, exactly.

Adding another thread where we sleep for `timeout-ms` and then call
`.destroyForcibly` on the process seems to do the trick.

Now, the only gotcha is if you're running an example and you Ctrl-C it,
the process might continue to hang around. I might fix that in the
future via capturing Ctrl-C and exiting cleanly, or something. But for
now, this is at least a substantial improvement.