it turns out that fair-queuing does work
I experimented a bit more with the mtserver example. I think this is
just a timing issue. When I increased the time it takes for a worker to
process a message from 1 second to 3 seconds, I started seeing the other
workers being used to.

I think what's happening is that the client scripts are all taking
slightly different amounts of time to start (Clojure has a large startup
cost), so by the time they're all running and the workers are processing
messages, it just happens to work out that a new client request is
available each time the worker finishes with the previous one.

Still seems like a suspicious coincidence, but the fact that the queuing
behavior is observable when I increase the time it takes for a worker to
process a task is encouraging.

This might also reinforce my comments about the zguide going on to
describe better patterns. I think it's the case that this behavior where
one worker can end up taking all the requests is an inherent weakness of
this pattern.